+++ projectroles/views.py
@@ -56,7 +56,7 @@ class ProjectPermissionObjectMixin(PermissionRequiredMixin):
     """Mixin for providing a Project object for permission checking"""
     def get_permission_object(self):
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -85,7 +85,7 @@ class RolePermissionMixin(LoggedInPermissionMixin):
     def has_permission(self):
         """Override has_permission to check perms depending on role"""
         try:
-            obj = RoleAssignment.objects.get(pk=self.kwargs['pk'])
+            obj = RoleAssignment.objects.get(omics_uuid=self.kwargs['role'])
 
             if obj.role.name == PROJECT_ROLE_OWNER:
                 # Modifying the project owner is not allowed in role views
@@ -149,7 +149,7 @@ class ProjectContextMixin(HTTPRefererMixin, ContextMixin):
         elif 'project' in self.kwargs:
             try:
                 context['project'] = Project.objects.get(
-                    pk=self.kwargs['project'])
+                    omics_uuid=self.kwargs['project'])
 
             except Project.DoesNotExist:
                 pass
@@ -253,6 +253,8 @@ class ProjectDetailView(
     """Project details view"""
     permission_required = 'projectroles.view_project'
     model = Project
+    slug_url_kwarg = 'project'
+    slug_field = 'omics_uuid'
 
     def get_context_data(self, *args, **kwargs):
         context = super(ProjectDetailView, self).get_context_data(
@@ -514,7 +516,7 @@ class ProjectModifyMixin(ModelFormMixin):
 
         messages.success(self.request, '{} {}d.'.format(type_str, form_action))
         return HttpResponseRedirect(reverse(
-            'project_detail', kwargs={'pk': project.pk}))
+            'project_detail', kwargs={'project': project.omics_uuid}))
 
 
 class ProjectCreateView(
@@ -530,7 +532,7 @@ class ProjectCreateView(
         parent"""
         if 'project' in self.kwargs:
             try:
-                obj = Project.objects.get(pk=self.kwargs['project'])
+                obj = Project.objects.get(omics_uuid=self.kwargs['project'])
                 return obj
 
             except Project.DoesNotExist:
@@ -541,7 +543,8 @@ class ProjectCreateView(
             *args, **kwargs)
 
         if 'project' in self.kwargs:
-            context['parent'] = Project.objects.get(pk=self.kwargs['project'])
+            context['parent'] = Project.objects.get(
+                omics_uuid=self.kwargs['project'])
 
         return context
 
@@ -555,15 +558,15 @@ class ProjectCreateView(
     def get(self, request, *args, **kwargs):
         """Override get() to limit project creation under other projects"""
         if 'project' in self.kwargs:
-            project = Project.objects.get(pk=self.kwargs['project'])
+            project = Project.objects.get(omics_uuid=self.kwargs['project'])
 
             if project.type != PROJECT_TYPE_CATEGORY:
                 messages.error(
                     self.request,
                     'Creating a project within a project is not allowed')
-                return HttpResponseRedirect(
-                    reverse(
-                        'project_detail', kwargs={'pk': project.pk}))
+                return HttpResponseRedirect(reverse(
+                    'project_detail',
+                    kwargs={'project': project.omics_uuid}))
 
         return super(ProjectCreateView, self).get(request, *args, **kwargs)
 
@@ -575,6 +578,8 @@ class ProjectUpdateView(
     permission_required = 'projectroles.update_project'
     model = Project
     form_class = ProjectForm
+    slug_url_kwarg = 'project'
+    slug_field = 'omics_uuid'
 
     def get_form_kwargs(self):
         kwargs = super(ProjectUpdateView, self).get_form_kwargs()
@@ -587,7 +592,7 @@ class ProjectUpdateView(
 
 class ProjectRoleView(
         LoginRequiredMixin, LoggedInPermissionMixin, ProjectContextMixin,
-        SingleObjectMixin, TemplateView):
+        TemplateView):
     """View for displaying project roles"""
     permission_required = 'projectroles.view_project_roles'
     template_name = 'projectroles/project_roles.html'
@@ -597,7 +602,7 @@ class ProjectRoleView(
         """Override get_object to provide a Project object for both template
         and permission checking"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['pk'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -606,19 +611,11 @@ class ProjectRoleView(
     def get_context_data(self, *args, **kwargs):
         context = super(ProjectRoleView, self).get_context_data(
             *args, **kwargs)
-
-        context['owner'] = self.object.get_owner()
-        context['delegate'] = self.object.get_delegate()
-        context['staff'] = self.object.get_staff()
-        context['members'] = self.object.get_members()
+        context['owner'] = context['project'].get_owner()
+        context['delegate'] = context['project'].get_delegate()
+        context['staff'] = context['project'].get_staff()
+        context['members'] = context['project'].get_members()
         return context
-
-    def dispatch(self, request, *args, **kwargs):
-        """Override dispatch to ensure self.object is provided to template"""
-        self.object = self.get_object()
-
-        return super(
-            ProjectRoleView, self).dispatch(request, *args, **kwargs)
 
 
 class RoleAssignmentModifyMixin(ModelFormMixin):
@@ -627,7 +624,7 @@ class RoleAssignmentModifyMixin(ModelFormMixin):
             *args, **kwargs)
 
         change_type = self.request.resolver_match.url_name.split('_')[1]
-        project = Project.objects.get(pk=self.kwargs['project'])
+        project = Project.objects.get(omics_uuid=self.kwargs['project'])
 
         if change_type != 'delete':
             context['preview_subject'] = get_role_change_subject(
@@ -640,7 +637,8 @@ class RoleAssignmentModifyMixin(ModelFormMixin):
                 role_name='{role_name}',
                 project_url=self.request.build_absolute_uri(reverse(
                     'project_detail',
-                    kwargs={'pk': project.pk}))).replace('\n', '\\n')
+                    kwargs={
+                        'project': project.omics_uuid}))).replace('\n', '\\n')
 
         return context
 
@@ -696,8 +694,8 @@ class RoleAssignmentModifyMixin(ModelFormMixin):
                     tl_event.set_status('FAILED', str(ex))
 
                 messages.error(self.request, str(ex))
-                return redirect(
-                    reverse('project_roles', kwargs={'pk': project.pk}))
+                return redirect(reverse(
+                    'project_roles', kwargs={'project': project.omics_uuid}))
 
             # Get object
             self.object = RoleAssignment.objects.get(
@@ -732,7 +730,9 @@ class RoleAssignmentModifyMixin(ModelFormMixin):
                 self.object.user.username,
                 self.object.role.name))
         return redirect(
-            reverse('project_roles', kwargs={'pk': self.object.project.pk}))
+            reverse(
+                'project_roles',
+                kwargs={'project': self.object.project.omics_uuid}))
 
 
 class RoleAssignmentCreateView(
@@ -746,7 +746,7 @@ class RoleAssignmentCreateView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -766,6 +766,8 @@ class RoleAssignmentUpdateView(
     """RoleAssignment updating view"""
     model = RoleAssignment
     form_class = RoleAssignmentForm
+    slug_url_kwarg = 'role'
+    slug_field = 'omics_uuid'
 
     def get_form_kwargs(self):
         """Pass current user to form"""
@@ -779,13 +781,15 @@ class RoleAssignmentDeleteView(
         DeleteView):
     """RoleAssignment deletion view"""
     model = RoleAssignment
+    slug_url_kwarg = 'role'
+    slug_field = 'omics_uuid'
 
     def post(self, *args, **kwargs):
         timeline = get_backend_api('timeline_backend')
         taskflow = get_backend_api('taskflow')
         tl_event = None
 
-        self.object = RoleAssignment.objects.get(pk=kwargs['pk'])
+        self.object = RoleAssignment.objects.get(omics_uuid=kwargs['role'])
         project = self.object.project
         user = self.object.user
         role = self.object.role
@@ -828,8 +832,9 @@ class RoleAssignmentDeleteView(
                     tl_event.set_status('FAILED', str(ex))
 
                 messages.error(self.request, str(ex))
-                return HttpResponseRedirect(redirect(
-                    reverse('project_roles', kwargs={'pk': project.pk})))
+                return HttpResponseRedirect(redirect(reverse(
+                    'project_roles',
+                    kwargs={'project': project.omics_uuid})))
 
         # Local save without Taskflow
         else:
@@ -850,7 +855,7 @@ class RoleAssignmentDeleteView(
                 user.username))
 
         return HttpResponseRedirect(reverse(
-            'project_roles', kwargs={'pk': project.pk}))
+            'project_roles', kwargs={'project': project.omics_uuid}))
 
     def get_form_kwargs(self):
         """Pass current user to form"""
@@ -871,7 +876,7 @@ class RoleAssignmentImportView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -883,7 +888,8 @@ class RoleAssignmentImportView(
 
         if request.user.is_superuser:
             projects = Project.objects.filter(
-                type=PROJECT_TYPE_PROJECT).exclude(pk=self.kwargs['project'])
+                type=PROJECT_TYPE_PROJECT).exclude(
+                omics_uuid=self.kwargs['project'])
 
             context['owned_projects'] = sorted(
                 [p for p in projects],
@@ -894,7 +900,7 @@ class RoleAssignmentImportView(
                 project__type=PROJECT_TYPE_PROJECT,
                 user=self.request.user,
                 role__name=PROJECT_ROLE_OWNER).exclude(
-                    project__pk=self.kwargs['project'])
+                    project__omics_uuid=self.kwargs['project'])
 
             if assignments.count() > 0:
                 context['owned_projects'] = sorted(
@@ -902,7 +908,7 @@ class RoleAssignmentImportView(
                     key=lambda x: x.get_full_title())
 
         context['previous_page'] = reverse(
-            'project_roles', kwargs={'pk': self.kwargs['project']})
+            'project_roles', kwargs={'project': self.kwargs['project']})
 
         return super(TemplateView, self).render_to_response(context)
 
@@ -915,7 +921,8 @@ class RoleAssignmentImportView(
             'import-mode' in post_data else None
 
         dest_project = self.get_permission_object()
-        source_project = Project.objects.get(pk=post_data['source-project'])
+        source_project = Project.objects.get(
+            omics_uuid=post_data['source-project'])
 
         ######################
         # Confirmation needed
@@ -946,7 +953,7 @@ class RoleAssignmentImportView(
                         user__in=import_users).order_by('user__username')
 
             context['previous_page'] = reverse(
-                'role_import', kwargs={'project': dest_project.pk})
+                'role_import', kwargs={'project': dest_project.omics_uuid})
 
             return super(TemplateView, self).render_to_response(context)
 
@@ -956,7 +963,7 @@ class RoleAssignmentImportView(
 
         import_keys = [
             key for key, val in post_data.items()
-            if key.startswith('import-field') and val == '1']
+            if key.startswith('import_field') and val == '1']
         import_count = len(import_keys)
 
         # Import/update
@@ -964,7 +971,8 @@ class RoleAssignmentImportView(
             import_users = []
 
             for key in import_keys:
-                source_as = RoleAssignment.objects.get(pk=key.split('-')[2])
+                source_as = RoleAssignment.objects.get(
+                    omics_uuid=key.split('_')[2])
 
                 try:
                     old_as = RoleAssignment.objects.get(
@@ -1052,13 +1060,13 @@ class RoleAssignmentImportView(
         # Delete
         del_keys = [
             key for key, val in post_data.items()
-            if key.startswith('delete-field') and val == '1']
+            if key.startswith('delete_field') and val == '1']
         del_count = len(del_keys)
 
         if import_mode == 'replace' and del_count > 0:
-            del_pks = [k.split('-')[2] for k in del_keys]
+            del_uuids = [k.split('_')[2] for k in del_keys]
             del_assignments = RoleAssignment.objects.filter(
-                pk__in=del_pks).order_by('user__username')
+                omics_uuid__in=del_uuids).order_by('user__username')
             del_users = [a.user for a in del_assignments]
 
             del_assignments.delete()
@@ -1107,7 +1115,7 @@ class RoleAssignmentImportView(
                     import_mode))
 
         return redirect(reverse(
-            'project_roles', kwargs={'pk': dest_project.pk}))
+            'project_roles', kwargs={'project': dest_project.omics_uuid}))
 
 
 # ProjectInvite Views ----------------------------------------------------
@@ -1115,7 +1123,7 @@ class RoleAssignmentImportView(
 
 class ProjectInviteView(
         LoginRequiredMixin, LoggedInPermissionMixin, ProjectContextMixin,
-        SingleObjectMixin, TemplateView):
+        TemplateView):
     """View for displaying and modifying project invites"""
     permission_required = 'projectroles.invite_users'
     template_name = 'projectroles/project_invites.html'
@@ -1125,7 +1133,7 @@ class ProjectInviteView(
         """Override get_object to provide a Project object for both template
         and permission checking"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -1136,18 +1144,11 @@ class ProjectInviteView(
             *args, **kwargs)
 
         context['invites'] = ProjectInvite.objects.filter(
-            project=self.kwargs['project'],
+            project=context['project'],
             active=True,
             date_expire__gt=timezone.now())
 
         return context
-
-    def dispatch(self, request, *args, **kwargs):
-        """Override dispatch to ensure self.object is provided to template"""
-        self.object = self.get_object()
-
-        return super(
-            ProjectInviteView, self).dispatch(request, *args, **kwargs)
 
 
 class ProjectInviteCreateView(
@@ -1161,7 +1162,7 @@ class ProjectInviteCreateView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -1190,7 +1191,7 @@ class ProjectInviteCreateView(
         """Pass current user to form"""
         kwargs = super(ProjectInviteCreateView, self).get_form_kwargs()
         kwargs.update({'current_user': self.request.user})
-        kwargs.update({'project': self.get_permission_object().pk})
+        kwargs.update({'project': self.get_permission_object().omics_uuid})
         return kwargs
 
     def form_valid(self, form):
@@ -1224,7 +1225,7 @@ class ProjectInviteCreateView(
         return redirect(
             reverse(
                 'role_invites',
-                kwargs={'project': self.object.project.pk}))
+                kwargs={'project': self.object.project.omics_uuid}))
 
 
 class ProjectInviteAcceptView(
@@ -1275,7 +1276,8 @@ class ProjectInviteAcceptView(
                 failed=True,
                 fail_desc='User already has roles in project')
             return redirect(reverse(
-                'project_detail', kwargs={'pk': invite.project.pk}))
+                'project_detail',
+                kwargs={'project': invite.project.omics_uuid}))
 
         except RoleAssignment.DoesNotExist:
             pass
@@ -1365,7 +1367,7 @@ class ProjectInviteAcceptView(
                 invite.project.title,
                 invite.role.name))
         return redirect(reverse(
-            'project_detail', kwargs={'pk': invite.project.pk}))
+            'project_detail', kwargs={'project': invite.project.omics_uuid}))
 
 
 class ProjectInviteResendView(
@@ -1376,7 +1378,7 @@ class ProjectInviteResendView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -1387,7 +1389,7 @@ class ProjectInviteResendView(
 
         try:
             invite = ProjectInvite.objects.get(
-                pk=self.kwargs['pk'],
+                omics_uuid=self.kwargs['invite'],
                 active=True)
 
         except ProjectInvite.DoesNotExist:
@@ -1424,7 +1426,7 @@ class ProjectInviteResendView(
                 localtime(invite.date_expire).strftime('%Y-%m-%d %H:%M')))
 
         return redirect(reverse(
-            'role_invites', kwargs={'project': invite.project.pk}))
+            'role_invites', kwargs={'project': invite.project.omics_uuid}))
 
 
 class ProjectInviteRevokeView(
@@ -1437,7 +1439,7 @@ class ProjectInviteRevokeView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['project'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist:
@@ -1450,15 +1452,15 @@ class ProjectInviteRevokeView(
         if 'project' in self.kwargs:
             try:
                 context['project'] = Project.objects.get(
-                    pk=self.kwargs['project'])
+                    omics_uuid=self.kwargs['project'])
 
             except Project.DoesNotExist:
                 pass
 
-        if 'pk' in self.kwargs:
+        if 'invite' in self.kwargs:
             try:
                 context['invite'] = ProjectInvite.objects.get(
-                    pk=self.kwargs['pk'])
+                    omics_uuid=self.kwargs['invite'])
 
             except ProjectInvite.DoesNotExist:
                 pass
@@ -1472,7 +1474,7 @@ class ProjectInviteRevokeView(
 
         try:
             invite = ProjectInvite.objects.get(
-                pk=kwargs['pk'])
+                omics_uuid=kwargs['invite'])
 
             invite.active = False
             invite.save()
@@ -1484,7 +1486,7 @@ class ProjectInviteRevokeView(
         # Add event in Timeline
         if timeline:
             timeline.add_event(
-                project=Project.objects.get(pk=self.kwargs['project']),
+                project=Project.objects.get(omics_uuid=self.kwargs['project']),
                 app_name=APP_NAME,
                 user=self.request.user,
                 event_name='invite_revoke',
@@ -1508,7 +1510,7 @@ class ProjectStarringAPIView(
     def get_permission_object(self):
         """Override get_permission_object for checking Project permission"""
         try:
-            obj = Project.objects.get(pk=self.kwargs['pk'])
+            obj = Project.objects.get(omics_uuid=self.kwargs['project'])
             return obj
 
         except Project.DoesNotExist: