+++ samplesheets/urls.py
@@ -4,27 +4,27 @@ from . import views
 
 urlpatterns = [
     url(
-        regex=r'^(?P<project>\d+)$',
+        regex=r'^(?P<project>[\w-]+)$',
         view=views.ProjectSheetsView.as_view(),
         name='project_sheets',
     ),
     url(
-        regex=r'^(?P<project>\d+)/study/(?P<study>\d+)$',
+        regex=r'^(?P<project>[\w-]+)/study/(?P<study>[\w-]+)$',
         view=views.ProjectSheetsView.as_view(),
         name='project_sheets',
     ),
     url(
-        regex=r'^(?P<project>\d+)/meta/(?P<subpage>[^\0]{0,256})$',
+        regex=r'^(?P<project>[\w-]+)/meta/(?P<subpage>[\w-]+)$',
         view=views.ProjectSheetsView.as_view(),
         name='project_sheets',
     ),
     url(
-        regex=r'^(?P<project>\d+)/import$',
+        regex=r'^(?P<project>[\w-]+)/import$',
         view=views.SampleSheetImportView.as_view(),
         name='sheet_import',
     ),
     url(
-        regex=r'^(?P<project>\d+)/delete$',
+        regex=r'^(?P<project>[\w-]+)/delete$',
         view=views.SampleSheetDeleteView.as_view(),
         name='sheet_delete',
     ),