+++ timeline/plugins.py
@@ -48,7 +48,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     #: FontAwesome icon ID string
     icon = 'clock-o'
 
-    #: Entry point URL ID (must take project pk as "project" argument)
+    #: Entry point URL ID (must take project omics_uuid as "project" argument)
     entry_point_url_id = 'project_timeline'
 
     #: Description string