+++ timeline/urls.py
@@ -4,13 +4,13 @@ from . import views
 
 urlpatterns = [
     url(
-        regex=r'^(?P<project>\d+)$',
+        regex=r'^(?P<project>[0-9a-f-]+)$',
         view=views.ProjectTimelineView.as_view(),
         name='project_timeline',
     ),
     url(
-        regex=r'^(?P<project>\d+)/(?P<object_model>[^\0]{0,256})/'
-              r'(?P<object_pk>\d+)$',
+        regex=r'^(?P<project>[0-9a-f-]+)/(?P<object_model>[\w-]+)/'
+              r'(?P<object_uuid>[\w-]+)$',
         view=views.ObjectTimelineView.as_view(),
         name='object_timeline',
     ),