+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - **General**
     - Graph output with pygraphviz for local development
     - ``ProjectPermissionObjectMixin`` for project apps
+    - The ``omics_uuid`` field to User model (#97)
 - **Projectroles**
     - Add ``get_project_link()`` in templatetags
 - **Samplesheets**
@@ -32,11 +33,15 @@ Changed
 - **General**
     - Upgrade site to Django 1.11.11
     - Upgrade site to Boostrap 4.0.0 Stable (#78)
+    - Use ``omics_uuid`` instead of ``pk`` in URLs and templates (#97)
     - Modify subtitle and page content containers for all apps
     - Site-wide CSS tweaks
     - Rename ``details_position`` to ``plugin_ordering`` in plugins (#90)
 - **Projectroles**
     - Search view improvements
+    - Refactor roles and invites views
+- **Timeline**
+    - Use ``omics_uuid`` for object lookup in ``plugins.get_object_link()`` (#97)
 
 Fixed
 -----